requirements.txt
mcp[cli]
python-pptx
Pillow
fonttools
boto3
openpyxl
python-dotenv